requirements_dev.txt
-r requirements.txt

flake8==3.5.0
mypy==0.560
bandit==1.4.0

pytest==3.0.7
pytest-asyncio==0.8.0
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
pytest-sugar==0.9.0
coverage==4.4.2
docker==2.7.0

Sphinx==1.6.6
sphinx-rtd-theme==0.2.4
watchdog==0.8.3

pip==9.0.1
wheel
tox==2.9.1

async-generator==1.8
async_timeout==2.0.0
cryptography==2.1.4
PyYAML==3.12

